refactor: 기존 SectionService에 정의된 createSection(addSection), deleteSection 메서드를 LineService로 이동
